chore(ggdeals): Reorganized views and viewmodels
