--- Library/GGDealsWishlist/GGDealsWishlistSettingsView.xaml.cs
@@ -1,12 +1,0 @@
-﻿using System.Windows.Controls;
-
-namespace GGDealsWishlist
-{
-    public partial class GGDealsWishlistSettingsView : UserControl
-    {
-        public GGDealsWishlistSettingsView()
-        {
-            InitializeComponent();
-        }
-    }
-}
