+++ Library/GGDealsWishlist/Views/SettingsView.xaml
@@ -1,11 +1,11 @@
-﻿<UserControl x:Class="GGDealsWishlist.GGDealsWishlistSettingsView"
+﻿<UserControl x:Class="GGDealsWishlist.Views.SettingsView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             xmlns:vm="clr-namespace:GGDealsWishlist"
+             xmlns:vm="clr-namespace:GGDealsWishlist.ViewModels"
              xmlns:c="clr-namespace:KNARZhelper.Controls"
-             d:DataContext="{d:DesignInstance Type=vm:GGDealsWishlistSettingsViewModel}"
+             d:DataContext="{d:DesignInstance Type=vm:SettingsViewModel}"
              mc:Ignorable="d"
              d:Background="LightGray"
              Padding="15, 10"
